CLN/TST: Clean groupby.test_allowlist
